Fixes after moving modules
